@@ -23,10 +23,10 @@ group= "com.theprogrammingturkey.chancecubes"
 archivesBaseName= "ChanceCubes"
 version= "${mc_version}-${mod_version}.${buildnumber}"
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(16)
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 minecraft {
-    mappings channel: 'official', version: '1.17.1'
+    mappings channel: 'official', version: '1.18.1'
 
     //accessTransformer = file('my_at.cfg')
 
@@ -34,7 +34,7 @@ minecraft {
     // these can be tweaked, removed, or duplicated as needed.
     runs {
         client {
-            workingDirectory project.file('run-1.17')
+            workingDirectory project.file('run-1.18')
 
             // Recommended logging data for a userdev environment
             property 'forge.logging.markers', ''
@@ -50,7 +50,7 @@ minecraft {
         }
 
         server {
-            workingDirectory project.file('run-server-1.17')
+            workingDirectory project.file('run-server-1.18')
 
             // Recommended logging data for a userdev environment
             property 'forge.logging.markers', ''
@@ -66,7 +66,7 @@ minecraft {
         }
 
         data {
-            workingDirectory project.file('run-1.17')
+            workingDirectory project.file('run-1.18')
 
             // Recommended logging data for a userdev environment
             property 'forge.logging.markers', ''
@@ -107,21 +107,3 @@ jar {
                     "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")],)
     }
 }
-
-//def reobfFile = file("$buildDir/reobfJar/output.jar")
-//def reobfArtifact = artifacts.add('default', reobfFile) {
-//    type 'jar'
-//    builtBy 'reobfJar'
-//}
-//publishing {
-//    publications {
-//        mavenJava(MavenPublication) {
-//            artifact reobfArtifact
-//        }
-//    }
-//    repositories {
-//        maven {
-//            url "file:///${project.projectDir}/mcmodsrepo"
-//        }
-//    }
-//}
