@@ -2,5 +2,5 @@ package chanceCubes.util;
 
 public enum GameTurn
 {
-	CPU, PLAYER;
+	CPU, PLAYER
 }
